Delete unused CSS files (#1594)
We currently get `jquery.jqplot` via npm, so it is no longer necessary
to ship our own copy of the file. `tabs_ie.css` and
`animated-overlay.gif` appear to be completely unused and have been
deleted.

There are a handful of other CSS files which may not be used, but I am
not confident enough to delete them at this time.