Eliminate GetUsersFromProject function
When there are no users linked to a project coverage file, the
GetUsersFromProject function returns false, assuming that an empty
query_result means an error.  Returning false created a 500 error
when the caller function attempted to loop over the return.

Since it is called in one place only, remove the function and set the DB
call in the controller directly.