Demote severity of 'No note text' to info
We've noticed tons of entries for this log message on open.cdash.org.
They stem from projects that choose to upload lots of logs for their builds,
some of which happen to be empty.

Since this is ultimately a client-side issue, demote the severity of this
log entry from 'error' to 'info' so as to make them easier to filter out on
the server side.