Fix handling of manageCoverage priority (#1617)
As is, the value of the priority in the dropdown in the table was used
to set the priority as the very last option, even if the priority was
changed using the "Change priority of selected files" capablities.

Change the logic to only use the prioritySelection if it is set and the
changePrioritySelected button was not pressed.

Fixes: #1616

---------

Co-authored-by: William Allen <16820599+williamjallen@users.noreply.github.com>