Pin PHP dependencies to most specific version possible (#1623)
We currently use range-based version specifiers in our `composer.json`.
To determine the specific version of a package we use, it is necessary
to check the autogenerated `composer.lock`.

Since CDash is a full software system instead of a library, there is no
need to pin our dependencies to a range of dependency versions. We only
test against the versions specified in `composer.lock`, so it is cleaner
to simply pin our dependencies to the most specific version possible and
rely upon Dependabot to keep `composer.json` and `composer.lock` synced
and up-to-date.

Other minor changes to `composer.json` in this PR:
- Moved `php-cs-fixer` to the list of dev dependencies
- Added the full list of PHP extensions we require
- Bumped the minimum PHP version to 8.1 since we use the `enum` keyword
which was first added in PHP 8.1

Fixes https://github.com/Kitware/CDash/issues/1313.