Treat autoremove builds query parameter as integer (#1627)
The query parameter for the limit clause in one of the build autoremoval
scripts is currently being passed to the database as a string. This is
problematic, and prints the following message to the logs:
`production.ERROR: You have an error in your SQL syntax; check the
manual that corresponds to your MySQL server version for the right
syntax to use near ''500'' at line 9`

MySQL does a limited amount of type conversion, but does not extend the
type conversion to limit clauses. By passing the query parameter as an
integer, the problem will be resolved.

I also addressed the only other case of this type of bug, even though
it's not directly related to the bug report per se.

This PR is a good example of why passing numeric strings around is a bad
idea...