Add missing final/abstract keywords (#1626)
The `FilterController` class was created before
https://github.com/Kitware/CDash/pull/1587 was merged, and is thus
missing the `final` designation.

`AbstractBuildController ` was missed amidst some logical conflicts when
previous PRs were merged and has been designated as `abstract`.

In the future, it might be interesting to create a PHPStan plugin which
ensures that all future controllers follow our naming and designation
conventions.