Delete unused dependencies: `predis` and `iron_mq` (#1632)
These two dependencies appear to be unused. We don't use Redis
currently, but we would have to re-add the Redis boilerplate if we were
to add it in the future. `iron_mq` is not likely to ever be used again.

I also cleaned up a couple minor issues introduced by #1623.