Migrate `manageBanner.php` from XSL to Blade (#1637)
This PR is part of an ongoing migration away from XSL. My goal is to be
able to eliminate XSL from the codebase entirely before the CDash 3.3
release.

While migrating the rendering engine from XSL to Blade, I performed a
significant refactor of the data marshaling logic for
`manageBanner.php`.