Remove unused build header view composer (#1651)
This PR removes a view composer which was rendered obsolete by #1351. In
the future, it might be reasonable to consider adding a view composer
which adds global information to the `cdash` view. That said, much of
the work of a view composer is already done by simply embedding required
information in our controller inheritance hierarchy, particularly by the
method utilized in #1483.