Add `view()` function to controllers (#1483)
We currently have a handful of things which get added to almost every
view. By adding a custom `view()` function in this PR, we can share more
code and reduce the number of locations which have to get updated if the
need for more "common" data arises.

I skipped over a few controllers which need more extensive work when
making these changes, and will fix those in a separate PR.