Turn on PHPStan "bleeding edge" (#1418)
PHPStan offers a set of "bleeding edge" rules which are the stable rules
which will be added in the next major release. We already use many of
these rules through our usage of various optional packages, but there's
no reason to not turn on additional static analysis rules. The only
major difference between bleeding edge and our current setup is that
bleeding edge mandates usage of the strict equality (`===` and `!==`)
operators. We already use the optional exception-checking rules which
are included in bleeding edge.

In addition to adding the bleeding edge rules, I have also enabled two
more optional rules which are not included in any of our optional
packages.