Migrate backend of `Project` model to Eloquent (#1556)
The `Project` model is one of the largest models in the CDash codebase,
and it would be completely impractical to convert the entire thing to
Eloquent at one time. Instead, I have taken an "adapter" approach
similar to the approach taken in #1539, in which the public-facing
interface of the old project class remains (mostly) the same, while the
business logic of most methods is outsourced to the new Eloquent-based
project model.

In addition to switching the code to use the Eloquent-based project when
possible, I have also undertaken a general refactor of the model,
deleting unused code, simplifying parameters and return types,
converting all database queries to Laravel's query system, and other
general maintenance tasks.

I have also completely eliminated the `projectrobot` table and
associated code since it is currently impossible to add new data to the
`projectrobot` table, and the existing data is not used for anything.

Given how large of a refactor this is, this PR should not be merged
until after the 3.2 release branch is created.