Add foreign key constraints for all projectids (#1655)
Our database schema is currently a mess. With no foreign keys defined,
it is easy to accidentally forget to update a relationship field
properly when an entity is created or deleted. Virtually all modern
relational databases support foreign key constraints, and there is no
reason for us to not add them.

In a first step towards #1442, this PR adds foreign key constraints to
all of the table columns which reference a project's ID. This should be
a relatively low risk operation for production systems, but it creates a
bit of a headache when writing tests, because creating any model which
references a project necessarily requires a project model to be created
as well.