Apply all "safe" PHP migration style rules (#1624)
PHP-CS-Fixer provides a ruleset for most major and minor PHP versions
which flags deprecated functions and ensures that our code style keeps
up with modern PHP style guidelines. This PR is part of a series of
improvements in which I intend to gradually turn on more style check
rules as we modernize the codebase.