Convert `viewCoverageFile.php` backend from XSL to Blade (#1664)
This PR continues our ongoing effort to move away from XSL as a
server-side templating engine.

I made a deliberate attempt to perform a one-to-one swap of the backend
rendering engine without touching the UI. This page has a variety of
confusing UI issues, and it is overdue for a major UI overhaul
eventually.