Refactor site-related routes (#1668)
`/viewSite.php` has been redirected to `/sites/<id>` and
`/siteStatistics.php` has been redirected to `/sites`.

As noted in a comment in the routes file, it would be nice to change the
behavior of `/sites` to show all sites the current user has access to,
rather than just limiting the route to admin users.