Finish transition from CDASH_USER_CREATE_PROJECTS (#1682)
Finish the replacement of the query for CDASH_USER_CREATE_PROJECTS
environment variable with the checking of the config singleton.

With the UserController transition, the link for project creation will
appear for each user when the variable is set.

All project creation by users is controlled by the `.env` variable
USER_CREATE_PROJECTS.

Finishes work started in  #1604