Add test for controller names (#1669)
CDash has a controller naming and designation convention, as discussed
in https://github.com/Kitware/CDash/pull/1587. Since then, I have
forgotten to mark several new controllers as "final". I have added a
test to enforce these rules automatically, rather than relying upon
human reviewers to catch it.

As seen in the diff for this PR, this test is quite effective. It caught
the fact that the controllers in the Auth directory were not marked with
the "final" keyword.