Migrate `BuildProperties` model backend to Eloquent (#1683)
This PR is part of our ongoing effort to migrate our models to Laravel's
Eloquent ORM library.

As noted in a comment, we should eventually consider whether the
`buildproperties` table should simply be a column on the `build` table ,
given that it has a one-to-one relationship with a corresponding build.
Performing such a refactor should be relatively straightforward.