Drop unused `random_compat` polyfill (#1694)
The `random_compat` polyfill is used to port a handful of functions
introduced in PHP 7 to PHP 5. Since we do not support PHP 5, we don't
need this dependency.

I have removed `random_compat` from our `composer.json`, but it is still
a required dependency for `league/oauth2-client` and is still in our
lock file.