Deprecate legacy `Config` class (#1698)
Everything the legacy `Config` class does can be done by an equivalent
Laravel helper function. As such, there is no longer a need for the
`Config` class, and it has been marked as deprecated. I tried to simply
replace each of the deprecated functions with laravel equivalents, but a
few quirks with our testing suite make it a nontrivial task.

I created a new function for getting the current CDash version in the
base controller class. This is now the preferred method to use for
getting the version. I intend to make a separate PR once this is merged
to simplify our versioning logic.