Convert backend of `Build` model to Eloquent (#1700)
This PR follows #1674 and converts the parts of the `Build` model which
interact directly with the `build` table to Eloquent wherever feasible.
I intend to slowly switch more parts of the codebase to the
Eloquent-based model, but that work will happen in separate PRs. By
simply having an Eloquent `Build` model present, we can do significantly
more in other Eloquent models.