Overhaul `BuildInformation` model (#1708)
All usages of the `BuildInformation` have been modified to use a new
Eloquent-based model.

In the future, it would be good to reconsider whether the
`buildinformation` table should be distinct from the `build` table as it
is now, or whether the columns of the `buildinformation` table should
simply be moved to the `build` table, given that there is a one-to-one
relationship between them.