Use Eloquent-based SubProject model whenever possible (#1709)
#1695 created an Eloquent-based `SubProject` model, but did not attempt
to change existing usages of the legacy model. This is a small follow-up
PR which expands the usage of the project-subproject relationship to
gain access to subprojects through the eloquent-based project model.
Amongst other things, this change should make a few pages slightly
faster by reducing the number of database queries executed when
iterating over subprojects. This PR also unifies the logic for
determining which subprojects were active on a given date. It is
probable that the prior inconsistent logic caused unexpected behavior in
rare cases.

I intend to work on subproject <-> build relationships in a future PR.