Add a Next Build table to the build summary page (#1697)
These changes add a "Next Build" table to go along with the "Current
Build" and "Previous Build" tables on the Build Summary page. The table
will be displayed if the next build is available, with the aim to
improve navigation between the various "build-specific" pages. Note that
the introduction of the third table prompted a reordering of the
existing tables so as to match the ordering in the "PREV | CURRENT |
NEXT" navigation bar at the top of the page.

---------

Co-authored-by: Shelly Belsky <sbelsky@salaxala.khq.kitware.com>