Round 2 of Kompose
With the last updates of the docker-compose system, give another go at
using the Kompose tool to create an openshift or kubernetes deployment
file.

Document the process of creating the input file and the command to
execute with Kompose.

The named files found here were created by using the command in the docs
with the addition of a `-o` flag to write the output to a single file
instead of a series of files as shown.  The YAML files listed in the
docs can be recreated with the content in this commit.