Bump PHPStan to level 8 (#1726)
PHPStan provides a set of [rule
levels](https://phpstan.org/user-guide/rule-levels) which provide
increasingly strict code analysis. We currently use level 6.

PHPStan level 8 checks nullable types in greater detail, which will
prevent a few common types of bugs, including the one fixed in #1725. I
was forced to also turn on level 7. Level 7 checks union types, which
will significantly improve our error detection abilities, but comes with
the caveat that checking for all of the error cases for functions which
return false instead of throwing an exception will cause additional
developer overhead.

Eventually, we should turn on level 9, but it would be pointless to do
so without turning on `treatPhpDocTypesAsCertain`. PHPStan cannot
currently understand Laravel's type casting system, which means that
we'd have to rely upon the PHPDoc types for level 9 to be of any
practical use.