Sanitize column names before `timeline.php` query (#1727)
`timeline.php` currently accepts arbitrary strings as column names,
which opens us up to SQL injection attacks. This PR sanitizes the column
names such that only the intended column names can be used. I also added
a basic test to ensure that this check continues to work in the future.
As part of adding the test, I was forced to migrate the API endpoint to
the new controller system without fully replacing the legacy controller.