Fix subproject dependency graph rendering (#1745)
### Current behavior
Currently, when loading the SubProject Dependencies page for a given
project, depending on the number of subprojects and the length of their
names, the rendering of the graph may overlap the text on the page.
<img width="959" alt="before_aligning"
src="https://github.com/Kitware/CDash/assets/71195502/402ec889-913d-48e4-8ebd-74506b0e051f">

Additionally, when hovering on top of nodes in the graph, the following
error is printed to the console:
<img width="467" alt="image"
src="https://github.com/Kitware/CDash/assets/71195502/9a475b24-9fc0-496e-9b2c-7dda26b8575f">

### New behavior
The graph will never overlap the text on the page anymore, and it set to
be rendered on the center of the page. The console error is fixed and
the tooltip now displays properly (the tiny white triangle connected to
the textbox that pops up on hover).

<img width="957" alt="after_aligning"
src="https://github.com/Kitware/CDash/assets/71195502/36462d92-18b5-4026-b72a-9530bab89114">

Co-authored-by: Shelly Belsky <sbelsky@salaxala.khq.kitware.com>