Rename central menu item from "current" to "latest" (#1730)
The navigation menu in the middle of the header menu is confusing,
because the "current" option takes the user to the most recent change,
while the "prev" and "next" options allow the user to navigate through
items one at a time. I changed the text from "current" to "latest" to
better reflect the action being performed, but I decided not to update
the underlying code to avoid breaking the API. I would like to hear any
feedback reviewers have regarding this decision.

After the change:

![image](https://github.com/Kitware/CDash/assets/16820599/1d46820d-029b-4f63-9270-2006b4a584a0)