Refactor error handler (#1762)
We currently simulate the behavior of the default Laravel error handler.
This approach doesn't handle some types of exceptions properly. This PR
introduces an updated error handler which modifies the result of the
error handler rather than attempting to simulate it.

The contents of this PR were originally part of #1742.