Fix Angular routing behavior after login (#1758)
Pages rendered by AngularJS which require authentication currently
redirect to the index page after login. This is inconsistent with the
behavior elsewhere, where the user is redirected back to the page they
attempted to access prior to being redirected to the login page.

This PR is a temporary patch to make the login redirect behavior more
consistent. The ultimate goal is to remove AngularJS entirely.