Delete invalid rows prior to creating FK constraint (#1768)
The CDash 3.3 release will include foreign-key constraints for the first
time. CDash does a poor job of cleaning up after itself when deleting
rows, so the new constraints will fail during the migration for many
existing CDash systems. The only way to add these constraints is delete
any data which violates them. This should not be problematic, because
there should be no way to access invalid data in the first place.

I also simplified the existing migrations due to an egregious violation
of the DRY principle...