Use Laravel storage mechanism for uploaded files
Remove all uses of CDASH_UPLOAD_DIRECTORY and replace them with
calls to Laravel's Storage facade.

Now that we are eliminating CDASH_UPLOAD_DIRECTORY, we can also
do away with the CheckDirectoryPermissions middleware, as its only
remaining purpose was to ensure that the web user had write access
to this directory.