Improve test label query performance (#1792)
`queryTests.php` is currently slow on some large production systems due
to the changes introduced in #1455. Adding this index improved the query
time by two orders of magnitude on a large production system. This PR
also fixes a small bug introduced by #1455, in which labels associated
with the same test run, but not the same build could appear in the
results for the wrong build.

I will backport this to the 3.2 release branch once it's merged into
master.