Migrate from Vue 2 to Vue 3 (#1794)
Vue 2 is [expected](https://v2.vuejs.org/lts/) to reach End-of-Life in
December 2023. Given that we plan to aggressively move our legacy
codebase to Vue in the coming months, it makes sense to perform this
upgrade now while the Vue portion of the codebase is still relatively
small.

This PR will also unblock a handful of other dependency upgrades.