Refactor legacy `User` model (#1802)
We currently have two user models: a legacy model, and an Eloquent
model. The reflection approach used to access legacy model members
through the Eloquent model is not conducive to static analysis, and is a
pain to work with. This PR moves a significant portion of the legacy
model logic to the Eloquent model. Additionally, I have improved the
User<->Password model relationship.