Convert backend of `Configure` model to Eloquent (#1815)
This PR continues our ongoing effort to use Eloquent for all database
interactions. I created two new Eloquent models, and refactored some
portions of the legacy BuildConfigure model to use them. The legacy
model in this case is difficult to work with, and I intend to return to
it at a future point for a more significant refactor.