Add JS linting to CI (#1827)
I made a mistake when creating #1374 by attempting to lint all of the
files while adding eslint. This iteration of my eslint work moves the
eslint step from our Mix build to a test which runs in our CI. I also
cleaned up the mix file at the same time, because it contained a bunch
of confusing JS code.

Once this PR is merged, I plan to lint sections of the codebase in
batches to slowly enable eslint. For now, most files are ignored.