Delete unused Laravel error SVGs (#1855)
Laravel provides a set of default backgrounds for various error pages.
We use custom error handlers for some errors, and the rest of the errors
do not use the Laravel backgrounds. The backgrounds are thus
unnecessary, and I have deleted them. If we want custom backgrounds for
error pages, we can create our CDash-themed SVGs.