Standardize usages of PHPUnit's `assertEquals()` (#1868)
According to [the PHPUnit docs for
`assertEquals()`](https://docs.phpunit.de/en/9.6/assertions.html#assertequals),
the first argument to the function should be the expected output, and
the second argument should be the actual value as produced by the test.
Currently, we have multiple places in the codebase where this ordering
of the arguments is not respected. Calling this function with the
arguments swapped can lead to confusion when debugging failed tests, as
PHPUnit's error messages present failed comparisons in terms of an
"expected" and an "actual" value.

The changes in this PR correct the instances where we use the wrong
ordering (except for the file `tests/Feature/Monitor.php` which is being
improved in PR https://github.com/Kitware/CDash/pull/1826).

There are many other PHPUnit assertion functions that follow this
argument format and are used in our test suites. Future PRs will correct
the cases for those functions where the argument ordering is not
respected as well.

It would also be nice if we could add tools to automatically enforce
this, in order to prevent this from happening in the future and improve
our development experience.