Clean up `UseCase` testing functionality (#1871)
Portions of the codebase are tested with heavy usage of mocking. As is
typical of tests which rely heavily upon mocking, many of the tests are
out of date. This PR is the first of several I intend to make to clean
up the `UseCase` portions of our test suite. This PR eliminates a large
amount of dead code, and adds type annotations to many of the models,
fixing more than 100 PHPStan errors.