Migrate the monitor page plot to d3 (#1826)
This PR introduces a [d3](https://d3js.org/) template for plotting line
graphs of time-series datasets, and converts the monitor page plot from
nvd3 to the new template. This change marks the first step of our
efforts to migrate to a single plotting/visualization dependency (see
https://github.com/Kitware/CDash/issues/1744). As we migrate more pages
to Vue in the future, other plots across the site will also be able to
use the new template and expand upon it.

_TODO: finalize style and appearance of plot_