Add unique constraint to email column of users table (#1899)
The email address column of the user table is considered to be unique
for login purposes, but the database allows multiple users to have the
same email address. This could potentially allow some users to be locked
out of their account. This PR adds a unique constraint to the email
column, so it is impossible to have multiple users with the same email.