Refactor Docker startup scripts (#1893)
Our Docker installation process was originally designed for testing and
development purposes. With Kubernetes and other container-based
deployment mechanisms becoming widespread, our Docker setup scripts are
in need of a refactor to make them deployment-quality.

Amongst a slew of general refactoring work, I made the following notable
changes:
- Apache is run under the current user by default, in case that user is
not www-data. This is particularly useful for Kubernetes deployments,
where the container is run under a random user.
- The Dockerfile has been moved to the root level of the repository and
renamed to `Dockerfile` in accordance with traditional Docker naming
conventions.
- `APP_KEY` is no longer generated by default. Users must specify it
manually, and the container with fail to start with a message showing a
random key if the `APP_KEY` environment variable is not set.
Automatically generating keys as was done before is a bad idea, because
the key (and thus, any data encrypted with it), will be lost if the
container is restarted for any reason.
- A root user is no longer created by default. This is particularly
useful for systems using OAuth or SAML, which should have all users
provided via the 3rd party authentication provider.
- Legacy configurations will no longer be migrated automatically, since
the migration will be lost if the container is restarted. If CDash
administrators wish to run the migration manually, they may do so by
running `php artisan config:migrate`.

I plan to make another PR to overhaul the Dockerfile and switch to
php-fpm sometime soon.