Add loading indicator Vue template (#1906)
We currently provide no information about the state of the page for
pages with API endpoints which take a long time to return. This PR adds
a Vue template which accepts a slot for content to be rendered when the
page is loaded, and a boolean prop which indicates whether the loading
indicator should be shown. I have initially used this template to add
loading indicators to two pages as an example for future use. I plan to
switch the loading indicator icon to a Font Awesome icon instead of a
grainy GIF once #1895 is merged.