Enforce a minimal `eslint` ignore list (#1935)
This PR adds a simple check to ensure all items in `.eslintignore` are
still valid paths. With our work on getting rid of technical debt, we
end up moving/deleting files quite a lot, and it's easy to forget to
edit the `eslint` ignore list as well. This new test will help us keep
this file up to date.