More gracefully handle misassigned configure records (#1940)
We encountered a situation where a parent build had a configure record
associated directly to it (and not to one of its children builds). In
this case, CDash would throw a 500 error when attempting to view the
configure output for this build.

While the cause was ultimately determined to be a non-standard XML
format, we should also prevent CDash from throwing an error in this
case, especially since the XML file in question was parsed
"successfully".