Enable Cypress component testing (#1941)
We currently use Cypress end-to-end testing to test entire pages. As we
move more of the UI to Vue, we will create a number of shared components
which are used on multiple pages. Cypress component testing allows us to
test these shared components in isolation by mounting each component
individually.