Bump CDash to v3.3 (#1967)
This PR should be the last thing merged before the CDash 3.3 branch is
created, so that it is the most recent commit on the 3.3 branch.

By merging this into master before creating a release branch for 3.3,
there will be less confusion about what the current version of master is
derived from. During the previous release cycle, master showed a version
of 3.1, while in reality it was dozens commits ahead of 3.2, which
confused the authors of multiple bug reports.