Add Vue component for standard tables (#1966)
This PR introduces a single standardized Vue component which can be used
for all tables across the site. This sortable table supports table cells
with text, links, and custom components. Column headers can be grouped.

As an example for developers to copy in the future, I have converted the
main table on the projects page to the new template.