Move static files to outer public directory (#1851)
CDash currently has a public directory which contains symlinks to the
legacy public directory. This prevents developers from building our
Docker image on Windows systems. Most places reference the new public
directory in the first place, so I have switched over the remaining
references to the old paths and deleted the symlinks.

I believe I was thorough and found all of the relevant locations, but
this is a difficult thing to be certain of and it is possible that I
missed something.