Add `createProject` GraphQL mutation (#2079)
This PR introduces the first mutation to the GraphQL API. This work can
be used as a template for future mutations. Future PRs will expand the
number of fields available in the project type, as well as introduce
project a project update mutation.

As part of this PR, I overhauled the project access control logic. It
now uses a Laravel Policy rather than a set of disorganized gates. I
left the legacy gates as-is, to be refactored at a future point in time
once the legacy project model becomes obsolete.