Replace most usages of deprecated PHPUnit `setMethods()` function  (#2086)
This PR is the second in the series for addressing changes that will
need to be made before https://github.com/Kitware/CDash/pull/1866 can be
merged. The PHPUnit function `setMethods()` is
[deprecated](https://github.com/sebastianbergmann/phpunit/blob/9.6/DEPRECATIONS.md#test-double-api),
so our usages of it were switched to `onlyMethods()` instead. The only
locations that weren't updated and still being suppressed in
`phpstan-baseline.neon` are:
* `app/cdash/include/Test/CDashTestCase.php`
* `app/cdash/include/Test/CDashUseCaseTestCase.php`
* `app/cdash/tests/case/CDash/Lib/Repository/GitHubTest.php`
* `app/cdash/tests/case/CDash/Model/BuildErrorTest.php`
* `tests/Unit/app/Validators/PasswordTest.php`

since a simple replacement of `setMethods()` by `onlyMethods()` does not
work for these cases.

Co-authored-by: William Allen <16820599+williamjallen@users.noreply.github.com>